@@ -11,25 +11,23 @@
   },
   "private": true,
   "dependencies": {
-    "@atlasmap/atlasmap.data.mapper": "^1.34.4",
+    "@atlasmap/atlasmap-data-mapper": "^1.35.0-SNAPSHOT",
     "is-primitive": "^3.0.0"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.5",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "tether": "1.4.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2",
-    "@angular/animations": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.6.3",
+    "@angular-devkit/build-ng-packagr": "~0.6.3",
+    "@angular/animations":"6.0.3",
+    "@angular/cli": "^6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/language-service": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/router": "6.0.3",
     "bootstrap": "^3.3.7",
     "chart.js": "^2.5.0",
     "classlist.js": "^1.1.20150312",
@@ -40,11 +38,19 @@
     "mkdirp": "^0.5.1",
     "moment": "^2.20.1",
     "ncp": "^2.0.0",
-    "ngx-bootstrap": "^2.0.2",
+    "ngx-bootstrap": "^3.0.0",
     "opn-cli": "^3.1.0",
-    "patternfly": "3.26.1",
+    "patternfly": "3.48.1",
     "patternfly-ng": "^3.0.0",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.2.0",
+    "tether": "1.4.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.7.2",
+    "zone.js": "^0.8.14",
+    "ng-packagr": "^3.0.0-rc.5",
+    "tsickle": ">=0.25.5",
+    "tslib":"^1.7.1",
+    "codelyzer": "~4.0.1",
+    "tslint": "~5.7.0"
   }
 }
