@@ -1,10 +1,10 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule,ChangeDetectorRef} from '@angular/core';
+import { NgModule, ChangeDetectorRef} from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+ import { HttpClientModule } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
 
-import { DataMapperModule } from '@atlasmap/atlasmap.data.mapper';
+import { DataMapperModule } from '@atlasmap/atlasmap-data-mapper';
 import { AtlasMapHostQEComponent } from './atlasmap.host.qe.component';
 
 @NgModule({
@@ -14,9 +14,9 @@ import { AtlasMapHostQEComponent } from './atlasmap.host.qe.component';
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     RouterModule.forRoot([]),
-    DataMapperModule
+    DataMapperModule.withInterceptor()
   ],
   exports: [
     AtlasMapHostQEComponent
